file read write test complete

+++ src/main/scala/org/beatific/people/utils/asyncfile/AsynchronousUnit.scala
@@ -4,8 +4,8 @@ import java.nio.channels.AsynchronousFileChannel
 
 class AsynchronousUnit(iter: AsynchronousIterator, key: Int) {
 
-  val begin: Long = iter.cap * key + iter.pos
-  val end: Long = if (iter.cap * (key + 1) - 1 + iter.pos > iter.max) iter.max - 1 else iter.cap * (key + 1) - 1 + iter.pos
+  val begin: Long = iter.cap.toLong * key.toLong + iter.pos
+  val end: Long = if (iter.cap.toLong * (key.toLong + 1L) - 1L + iter.pos > iter.max) iter.max - 1L else iter.cap.toLong * (key.toLong + 1L) - 1L + iter.pos
 
   def cache(temp: String) {
     iter.cache :+= temp
